Update github-assignment2
I wrote a static method which takes an Array of integers then established a loop to take the sum of all terms then divided to term number to find average. After that I used another loop to take the difference between the terms and the average. Then I used an array to store the values of differences and printed them.